major rewrite
but it's still compatible with 2.x.x config so only minor version increase
