+++ lib/config/getConfig.js
@@ -5,7 +5,7 @@ const getKey = (e, d) => e === undefined ? d : e
 module.exports = () => {
   var config
   try {
-    config = require('../config.json')
+    config = require('../../config.json')
     if (!config.discord.token || !config.facebook.username || !config.facebook.password) {
       log.warn('getConfig', 'Token/username/password not found, getting env config')
       config = getEnvConfig()
