+++ lib/config/getEnvConfig.js
@@ -8,8 +8,8 @@ module.exports = () => {
       showUsername: yn(process.env.FACEBOOK_SHOW_USERNAME),
       boldUsername: yn(process.env.FACEBOOK_BOLD_USERNAME),
       filter: {
-        whitelist: (process.env.FACEBOOK_FILTER_WHITELIST || '').split(','),
-        blacklist: (process.env.FACEBOOK_FILTER_BLACKLIST || '').split(',')
+        whitelist: (process.env.FACEBOOK_FILTER_WHITELIST || '').split(',').filter(el => el), // sort out empty elements
+        blacklist: (process.env.FACEBOOK_FILTER_BLACKLIST || '').split(',').filter(el => el) // sort out empty elements
       }
     },
     discord: {
@@ -22,7 +22,7 @@ module.exports = () => {
     logLevel: process.env.MISCORD_LOG_LEVEL
   }
   if (!config.discord.token || !config.facebook.username || !config.facebook.password) {
-    require('./error.js')('Env: token/username/password missing.')
+    require('../error.js')('Env: token/username/password missing.')
   }
   return config
 }
